requirements.txt
aiogram==3.13.1
supabase==2.4.6
python-dotenv==1.0.1
